+++ src/frontend/component/logo/logo.css
@@ -1,0 +1,12 @@
+.Logo {
+    text-align: center;
+    margin: auto;
+    padding: auto;
+    height: 15vh;
+    width: 40vw;
+}
+.img {
+    margin: auto;
+    height: 90%;
+    width: 80%;
+}
