+++ src/frontend/component/poster/corgiFull/corgiFull.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import corgifull from '../../../../assets/corgi-full.png';
+
+const CI = () => (
+    <div className="Corgi">
+        <img className="img" src={corgifull} />
+    </div>
+)
+
+export default CI
