+++ src/frontend/container/App.css
@@ -1,3 +1,4 @@
 .App {
+    margin-top: 5vh;
     display: grid;
 }
